Split Track.cpp and Track.h
