+++ include/AutomationPattern.h
@@ -31,6 +31,7 @@
 #include <QtCore/QPointer>
 
 #include "Track.h"
+#include "TrackContentObject.h"
 
 
 class AutomationTrack;
