+++ src/core/Track.cpp
@@ -40,2402 +40,22 @@
 #include <assert.h>
 #include <cstdlib>
 
-#include <QLayout>
-#include <QLinearGradient>
 #include <QMenu>
 #include <QMouseEvent>
 #include <QPainter>
-#include <QStyleOption>
 #include <QVariant>
-#include <QClipboard>
 
 
 #include "AutomationPattern.h"
 #include "AutomationTrack.h"
-#include "AutomationEditor.h"
-#include "BBEditor.h"
 #include "BBTrack.h"
 #include "BBTrackContainer.h"
 #include "ConfigManager.h"
-#include "Clipboard.h"
-#include "ColorChooser.h"
-#include "embed.h"
 #include "Engine.h"
-#include "GuiApplication.h"
-#include "FxMixerView.h"
-#include "gui_templates.h"
-#include "MainWindow.h"
-#include "Mixer.h"
-#include "ProjectJournal.h"
+#include "InstrumentTrack.h"
 #include "SampleTrack.h"
 #include "Song.h"
-#include "SongEditor.h"
-#include "StringPairDrag.h"
-#include "TextFloat.h"
 
-
-/*! The width of the resize grip in pixels
- */
-const int RESIZE_GRIP_WIDTH = 4;
-
-/*! Alternate between a darker and a lighter background color every 4 bars
- */
-const int BARS_PER_GROUP = 4;
-
-
-/*! A pointer for that text bubble used when moving segments, etc.
- *
- * In a number of situations, LMMS displays a floating text bubble
- * beside the cursor as you move or resize elements of a track about.
- * This pointer keeps track of it, as you only ever need one at a time.
- */
-TextFloat * TrackContentObjectView::s_textFloat = NULL;
-
-
-// ===========================================================================
-// TrackContentObject
-// ===========================================================================
-/*! \brief Create a new TrackContentObject
- *
- *  Creates a new track content object for the given track.
- *
- * \param _track The track that will contain the new object
- */
-TrackContentObject::TrackContentObject( Track * track ) :
-	Model( track ),
-	m_track( track ),
-	m_startPosition(),
-	m_length(),
-	m_mutedModel( false, this, tr( "Mute" ) ),
-	m_selectViewOnCreate( false ),
-	m_color( 128, 128, 128 ),
-	m_useCustomClipColor( false )
-{
-	if( getTrack() )
-	{
-		getTrack()->addTCO( this );
-	}
-	setJournalling( false );
-	movePosition( 0 );
-	changeLength( 0 );
-	setJournalling( true );
-}
-
-
-
-
-/*! \brief Destroy a TrackContentObject
- *
- *  Destroys the given track content object.
- *
- */
-TrackContentObject::~TrackContentObject()
-{
-	emit destroyedTCO();
-
-	if( getTrack() )
-	{
-		getTrack()->removeTCO( this );
-	}
-}
-
-
-
-
-/*! \brief Move this TrackContentObject's position in time
- *
- *  If the track content object has moved, update its position.  We
- *  also add a journal entry for undo and update the display.
- *
- * \param _pos The new position of the track content object.
- */
-void TrackContentObject::movePosition( const TimePos & pos )
-{
-	TimePos newPos = qMax(0, pos.getTicks());
-	if (m_startPosition != newPos)
-	{
-		Engine::mixer()->requestChangeInModel();
-		m_startPosition = newPos;
-		Engine::mixer()->doneChangeInModel();
-		Engine::getSong()->updateLength();
-		emit positionChanged();
-	}
-}
-
-
-
-
-/*! \brief Change the length of this TrackContentObject
- *
- *  If the track content object's length has changed, update it.  We
- *  also add a journal entry for undo and update the display.
- *
- * \param _length The new length of the track content object.
- */
-void TrackContentObject::changeLength( const TimePos & length )
-{
-	m_length = length;
-	Engine::getSong()->updateLength();
-	emit lengthChanged();
-}
-
-
-
-
-bool TrackContentObject::comparePosition(const TrackContentObject *a, const TrackContentObject *b)
-{
-	return a->startPosition() < b->startPosition();
-}
-
-
-
-
-/*! \brief Copies the state of a TrackContentObject to another TrackContentObject
- *
- *  This method copies the state of a TCO to another TCO
- */
-void TrackContentObject::copyStateTo( TrackContentObject *src, TrackContentObject *dst )
-{
-	// If the node names match we copy the state
-	if( src->nodeName() == dst->nodeName() ){
-		QDomDocument doc;
-		QDomElement parent = doc.createElement( "StateCopy" );
-		src->saveState( doc, parent );
-
-		const TimePos pos = dst->startPosition();
-		dst->restoreState( parent.firstChild().toElement() );
-		dst->movePosition( pos );
-
-		AutomationPattern::resolveAllIDs();
-		GuiApplication::instance()->automationEditor()->m_editor->updateAfterPatternChange();
-	}
-}
-
-
-
-
-/*! \brief Mutes this TrackContentObject
- *
- *  Restore the previous state of this track content object.  This will
- *  restore the position or the length of the track content object
- *  depending on what was changed.
- *
- * \param _je The journal entry to undo
- */
-void TrackContentObject::toggleMute()
-{
-	m_mutedModel.setValue( !m_mutedModel.value() );
-	emit dataChanged();
-}
-
-
-
-
-TimePos TrackContentObject::startTimeOffset() const
-{
-	return m_startTimeOffset;
-}
-
-
-
-
-void TrackContentObject::setStartTimeOffset( const TimePos &startTimeOffset )
-{
-	m_startTimeOffset = startTimeOffset;
-}
-
-// Update TCO color if it follows the track color
-void TrackContentObject::updateColor()
-{
-	if( ! m_useCustomClipColor )
-	{
-		emit trackColorChanged();
-	}
-}
-
-
-void TrackContentObject::useCustomClipColor( bool b )
-{
-	m_useCustomClipColor = b;
-	updateColor();
-}
-
-
-bool TrackContentObject::hasColor()
-{
-	return usesCustomClipColor() || getTrack()->useColor();
-}
-
-
-
-
-
-// ===========================================================================
-// trackContentObjectView
-// ===========================================================================
-/*! \brief Create a new trackContentObjectView
- *
- *  Creates a new track content object view for the given
- *  track content object in the given track view.
- *
- * \param _tco The track content object to be displayed
- * \param _tv  The track view that will contain the new object
- */
-TrackContentObjectView::TrackContentObjectView( TrackContentObject * tco,
-							TrackView * tv ) :
-	selectableObject( tv->getTrackContentWidget() ),
-	ModelView( NULL, this ),
-	m_tco( tco ),
-	m_trackView( tv ),
-	m_action( NoAction ),
-	m_initialMousePos( QPoint( 0, 0 ) ),
-	m_initialMouseGlobalPos( QPoint( 0, 0 ) ),
-	m_initialTCOPos( TimePos(0) ),
-	m_initialTCOEnd( TimePos(0) ),
-	m_initialOffsets( QVector<TimePos>() ),
-	m_hint( NULL ),
-	m_mutedColor( 0, 0, 0 ),
-	m_mutedBackgroundColor( 0, 0, 0 ),
-	m_selectedColor( 0, 0, 0 ),
-	m_textColor( 0, 0, 0 ),
-	m_textShadowColor( 0, 0, 0 ),
-	m_BBPatternBackground( 0, 0, 0 ),
-	m_gradient( true ),
-	m_mouseHotspotHand( 0, 0 ),
-	m_cursorSetYet( false ),
-	m_needsUpdate( true )
-{
-	if( s_textFloat == NULL )
-	{
-		s_textFloat = new TextFloat;
-		s_textFloat->setPixmap( embed::getIconPixmap( "clock" ) );
-	}
-
-	setAttribute( Qt::WA_OpaquePaintEvent, true );
-	setAttribute( Qt::WA_DeleteOnClose, true );
-	setFocusPolicy( Qt::StrongFocus );
-	setCursor( QCursor( embed::getIconPixmap( "hand" ), m_mouseHotspotHand.width(), m_mouseHotspotHand.height() ) );
-	move( 0, 0 );
-	show();
-
-	setFixedHeight( tv->getTrackContentWidget()->height() - 1);
-	setAcceptDrops( true );
-	setMouseTracking( true );
-
-	connect( m_tco, SIGNAL( lengthChanged() ),
-			this, SLOT( updateLength() ) );
-	connect( gui->songEditor()->m_editor->zoomingModel(), SIGNAL( dataChanged() ), this, SLOT( updateLength() ) );
-	connect( m_tco, SIGNAL( positionChanged() ),
-			this, SLOT( updatePosition() ) );
-	connect( m_tco, SIGNAL( destroyedTCO() ), this, SLOT( close() ) );
-	setModel( m_tco );
-	connect( m_tco, SIGNAL( trackColorChanged() ), this, SLOT( update() ) );
-	connect( m_trackView->getTrackOperationsWidget(), SIGNAL( colorParented() ), this, SLOT( useTrackColor() ) );
-
-	m_trackView->getTrackContentWidget()->addTCOView( this );
-	updateLength();
-	updatePosition();
-}
-
-
-
-
-/*! \brief Destroy a trackContentObjectView
- *
- *  Destroys the given track content object view.
- *
- */
-TrackContentObjectView::~TrackContentObjectView()
-{
-	delete m_hint;
-	// we have to give our track-container the focus because otherwise the
-	// op-buttons of our track-widgets could become focus and when the user
-	// presses space for playing song, just one of these buttons is pressed
-	// which results in unwanted effects
-	m_trackView->trackContainerView()->setFocus();
-}
-
-
-/*! \brief Update a TrackContentObjectView
- *
- *  TCO's get drawn only when needed,
- *  and when a TCO is updated,
- *  it needs to be redrawn.
- *
- */
-void TrackContentObjectView::update()
-{
-	if( !m_cursorSetYet )
-	{
-		setCursor( QCursor( embed::getIconPixmap( "hand" ), m_mouseHotspotHand.width(), m_mouseHotspotHand.height() ) );
-		m_cursorSetYet = true;
-	}
-
-	if( fixedTCOs() )
-	{
-		updateLength();
-	}
-	m_needsUpdate = true;
-	selectableObject::update();
-}
-
-
-
-/*! \brief Does this trackContentObjectView have a fixed TCO?
- *
- *  Returns whether the containing trackView has fixed
- *  TCOs.
- *
- * \todo What the hell is a TCO here - track content object?  And in
- *  what circumstance are they fixed?
- */
-bool TrackContentObjectView::fixedTCOs()
-{
-	return m_trackView->trackContainerView()->fixedTCOs();
-}
-
-
-
-// qproperty access functions, to be inherited & used by TCOviews
-//! \brief CSS theming qproperty access method
-QColor TrackContentObjectView::mutedColor() const
-{ return m_mutedColor; }
-
-QColor TrackContentObjectView::mutedBackgroundColor() const
-{ return m_mutedBackgroundColor; }
-
-QColor TrackContentObjectView::selectedColor() const
-{ return m_selectedColor; }
-
-QColor TrackContentObjectView::textColor() const
-{ return m_textColor; }
-
-QColor TrackContentObjectView::textBackgroundColor() const
-{
-	return m_textBackgroundColor;
-}
-
-QColor TrackContentObjectView::textShadowColor() const
-{ return m_textShadowColor; }
-
-QColor TrackContentObjectView::BBPatternBackground() const
-{ return m_BBPatternBackground; }
-
-bool TrackContentObjectView::gradient() const
-{ return m_gradient; }
-
-//! \brief CSS theming qproperty access method
-void TrackContentObjectView::setMutedColor( const QColor & c )
-{ m_mutedColor = QColor( c ); }
-
-void TrackContentObjectView::setMutedBackgroundColor( const QColor & c )
-{ m_mutedBackgroundColor = QColor( c ); }
-
-void TrackContentObjectView::setSelectedColor( const QColor & c )
-{ m_selectedColor = QColor( c ); }
-
-void TrackContentObjectView::setTextColor( const QColor & c )
-{ m_textColor = QColor( c ); }
-
-void TrackContentObjectView::setTextBackgroundColor( const QColor & c )
-{
-	m_textBackgroundColor = c;
-}
-
-void TrackContentObjectView::setTextShadowColor( const QColor & c )
-{ m_textShadowColor = QColor( c ); }
-
-void TrackContentObjectView::setBBPatternBackground( const QColor & c )
-{ m_BBPatternBackground = QColor( c ); }
-
-void TrackContentObjectView::setGradient( const bool & b )
-{ m_gradient = b; }
-
-void TrackContentObjectView::setMouseHotspotHand(const QSize & s)
-{
-	m_mouseHotspotHand = s;
-}
-
-// access needsUpdate member variable
-bool TrackContentObjectView::needsUpdate()
-{ return m_needsUpdate; }
-void TrackContentObjectView::setNeedsUpdate( bool b )
-{ m_needsUpdate = b; }
-
-/*! \brief Close a trackContentObjectView
- *
- *  Closes a track content object view by asking the track
- *  view to remove us and then asking the QWidget to close us.
- *
- * \return Boolean state of whether the QWidget was able to close.
- */
-bool TrackContentObjectView::close()
-{
-	m_trackView->getTrackContentWidget()->removeTCOView( this );
-	return QWidget::close();
-}
-
-
-
-
-/*! \brief Removes a trackContentObjectView from its track view.
- *
- *  Like the close() method, this asks the track view to remove this
- *  track content object view.  However, the track content object is
- *  scheduled for later deletion rather than closed immediately.
- *
- */
-void TrackContentObjectView::remove()
-{
-	m_trackView->getTrack()->addJournalCheckPoint();
-
-	// delete ourself
-	close();
-	m_tco->deleteLater();
-}
-
-
-
-
-/*! \brief Updates a trackContentObjectView's length
- *
- *  If this track content object view has a fixed TCO, then we must
- *  keep the width of our parent.  Otherwise, calculate our width from
- *  the track content object's length in pixels adding in the border.
- *
- */
-void TrackContentObjectView::updateLength()
-{
-	if( fixedTCOs() )
-	{
-		setFixedWidth( parentWidget()->width() );
-	}
-	else
-	{
-		setFixedWidth(
-		static_cast<int>( m_tco->length() * pixelsPerBar() /
-					TimePos::ticksPerBar() ) + 1 /*+
-						TCO_BORDER_WIDTH * 2-1*/ );
-	}
-	m_trackView->trackContainerView()->update();
-}
-
-
-
-
-/*! \brief Updates a trackContentObjectView's position.
- *
- *  Ask our track view to change our position.  Then make sure that the
- *  track view is updated in case this position has changed the track
- *  view's length.
- *
- */
-void TrackContentObjectView::updatePosition()
-{
-	m_trackView->getTrackContentWidget()->changePosition();
-	// moving a TCO can result in change of song-length etc.,
-	// therefore we update the track-container
-	m_trackView->trackContainerView()->update();
-}
-
-
-
-
-void TrackContentObjectView::changeClipColor()
-{
-	// Get a color from the user
-	QColor new_color = ColorChooser( this ).withPalette( ColorChooser::Palette::Track )->getColor( m_tco->color() );
-	if( ! new_color.isValid() )
-	{ return; }
-	
-	// Use that color
-	m_tco->setColor( new_color );
-	m_tco->useCustomClipColor( true );
-	update();
-}
-
-
-
-void TrackContentObjectView::useTrackColor()
-{
-	m_tco->useCustomClipColor( false );
-	update();
-}
-
-
-
-
-
-/*! \brief Change the trackContentObjectView's display when something
- *  being dragged enters it.
- *
- *  We need to notify Qt to change our display if something being
- *  dragged has entered our 'airspace'.
- *
- * \param dee The QDragEnterEvent to watch.
- */
-void TrackContentObjectView::dragEnterEvent( QDragEnterEvent * dee )
-{
-	TrackContentWidget * tcw = getTrackView()->getTrackContentWidget();
-	TimePos tcoPos = TimePos( m_tco->startPosition() );
-
-	if( tcw->canPasteSelection( tcoPos, dee ) == false )
-	{
-		dee->ignore();
-	}
-	else
-	{
-		StringPairDrag::processDragEnterEvent( dee, "tco_" +
-					QString::number( m_tco->getTrack()->type() ) );
-	}
-}
-
-
-
-
-/*! \brief Handle something being dropped on this trackContentObjectView.
- *
- *  When something has been dropped on this trackContentObjectView, and
- *  it's a track content object, then use an instance of our dataFile reader
- *  to take the xml of the track content object and turn it into something
- *  we can write over our current state.
- *
- * \param de The QDropEvent to handle.
- */
-void TrackContentObjectView::dropEvent( QDropEvent * de )
-{
-	QString type = StringPairDrag::decodeKey( de );
-	QString value = StringPairDrag::decodeValue( de );
-
-	// Track must be the same type to paste into
-	if( type != ( "tco_" + QString::number( m_tco->getTrack()->type() ) ) )
-	{
-		return;
-	}
-
-	// Defer to rubberband paste if we're in that mode
-	if( m_trackView->trackContainerView()->allowRubberband() == true )
-	{
-		TrackContentWidget * tcw = getTrackView()->getTrackContentWidget();
-		TimePos tcoPos = TimePos( m_tco->startPosition() );
-
-		if( tcw->pasteSelection( tcoPos, de ) == true )
-		{
-			de->accept();
-		}
-		return;
-	}
-
-	// Don't allow pasting a tco into itself.
-	QObject* qwSource = de->source();
-	if( qwSource != NULL &&
-	    dynamic_cast<TrackContentObjectView *>( qwSource ) == this )
-	{
-		return;
-	}
-
-	// Copy state into existing tco
-	DataFile dataFile( value.toUtf8() );
-	TimePos pos = m_tco->startPosition();
-	QDomElement tcos = dataFile.content().firstChildElement( "tcos" );
-	m_tco->restoreState( tcos.firstChildElement().firstChildElement() );
-	m_tco->movePosition( pos );
-	AutomationPattern::resolveAllIDs();
-	de->accept();
-}
-
-
-
-
-/*! \brief Handle a dragged selection leaving our 'airspace'.
- *
- * \param e The QEvent to watch.
- */
-void TrackContentObjectView::leaveEvent( QEvent * e )
-{
-	if( cursor().shape() != Qt::BitmapCursor )
-	{
-		setCursor( QCursor( embed::getIconPixmap( "hand" ), m_mouseHotspotHand.width(), m_mouseHotspotHand.height() ) );
-	}
-	if( e != NULL )
-	{
-		QWidget::leaveEvent( e );
-	}
-}
-
-/*! \brief Create a DataFile suitable for copying multiple trackContentObjects.
- *
- *	trackContentObjects in the vector are written to the "tcos" node in the
- *  DataFile.  The trackContentObjectView's initial mouse position is written
- *  to the "initialMouseX" node in the DataFile.  When dropped on a track,
- *  this is used to create copies of the TCOs.
- *
- * \param tcos The trackContectObjects to save in a DataFile
- */
-DataFile TrackContentObjectView::createTCODataFiles(
-    				const QVector<TrackContentObjectView *> & tcoViews) const
-{
-	Track * t = m_trackView->getTrack();
-	TrackContainer * tc = t->trackContainer();
-	DataFile dataFile( DataFile::DragNDropData );
-	QDomElement tcoParent = dataFile.createElement( "tcos" );
-
-	typedef QVector<TrackContentObjectView *> tcoViewVector;
-	for( tcoViewVector::const_iterator it = tcoViews.begin();
-			it != tcoViews.end(); ++it )
-	{
-		// Insert into the dom under the "tcos" element
-		Track* tcoTrack = ( *it )->m_trackView->getTrack();
-		int trackIndex = tc->tracks().indexOf( tcoTrack );
-		QDomElement tcoElement = dataFile.createElement( "tco" );
-		tcoElement.setAttribute( "trackIndex", trackIndex );
-		tcoElement.setAttribute( "trackType", tcoTrack->type() );
-		tcoElement.setAttribute( "trackName", tcoTrack->name() );
-		( *it )->m_tco->saveState( dataFile, tcoElement );
-		tcoParent.appendChild( tcoElement );
-	}
-
-	dataFile.content().appendChild( tcoParent );
-
-	// Add extra metadata needed for calculations later
-	int initialTrackIndex = tc->tracks().indexOf( t );
-	if( initialTrackIndex < 0 )
-	{
-		printf("Failed to find selected track in the TrackContainer.\n");
-		return dataFile;
-	}
-	QDomElement metadata = dataFile.createElement( "copyMetadata" );
-	// initialTrackIndex is the index of the track that was touched
-	metadata.setAttribute( "initialTrackIndex", initialTrackIndex );
-	metadata.setAttribute( "trackContainerId", tc->id() );
-	// grabbedTCOPos is the pos of the bar containing the TCO we grabbed
-	metadata.setAttribute( "grabbedTCOPos", m_tco->startPosition() );
-
-	dataFile.content().appendChild( metadata );
-
-	return dataFile;
-}
-
-void TrackContentObjectView::paintTextLabel(QString const & text, QPainter & painter)
-{
-	if (text.trimmed() == "")
-	{
-		return;
-	}
-
-	painter.setRenderHint( QPainter::TextAntialiasing );
-
-	QFont labelFont = this->font();
-	labelFont.setHintingPreference( QFont::PreferFullHinting );
-	painter.setFont( labelFont );
-
-	const int textTop = TCO_BORDER_WIDTH + 1;
-	const int textLeft = TCO_BORDER_WIDTH + 3;
-
-	QFontMetrics fontMetrics(labelFont);
-	QString elidedPatternName = fontMetrics.elidedText(text, Qt::ElideMiddle, width() - 2 * textLeft);
-
-	if (elidedPatternName.length() < 2)
-	{
-		elidedPatternName = text.trimmed();
-	}
-
-	painter.fillRect(QRect(0, 0, width(), fontMetrics.height() + 2 * textTop), textBackgroundColor());
-
-	int const finalTextTop = textTop + fontMetrics.ascent();
-	painter.setPen(textShadowColor());
-	painter.drawText( textLeft + 1, finalTextTop + 1, elidedPatternName );
-	painter.setPen( textColor() );
-	painter.drawText( textLeft, finalTextTop, elidedPatternName );
-}
-
-/*! \brief Handle a mouse press on this trackContentObjectView.
- *
- *  Handles the various ways in which a trackContentObjectView can be
- *  used with a click of a mouse button.
- *
- *  * If our container supports rubber band selection then handle
- *    selection events.
- *  * or if shift-left button, add this object to the selection
- *  * or if ctrl-left button, start a drag-copy event
- *  * or if just plain left button, resize if we're resizeable
- *  * or if ctrl-middle button, mute the track content object
- *  * or if middle button, maybe delete the track content object.
- *
- * \param me The QMouseEvent to handle.
- */
-void TrackContentObjectView::mousePressEvent( QMouseEvent * me )
-{
-	// Right now, active is only used on right/mid clicks actions, so we use a ternary operator
-	// to avoid the overhead of calling getClickedTCOs when it's not used
-	auto active = me->button() == Qt::LeftButton
-		? QVector<TrackContentObjectView *>()
-		: getClickedTCOs();
-
-	setInitialPos( me->pos() );
-	setInitialOffsets();
-	if( !fixedTCOs() && me->button() == Qt::LeftButton )
-	{
-		if( me->modifiers() & Qt::ControlModifier )
-		{
-			if( isSelected() )
-			{
-				m_action = CopySelection;
-			}
-			else
-			{
-				m_action = ToggleSelected;
-			}
-		}
-		else if( !me->modifiers()
-			|| (me->modifiers() & Qt::AltModifier)
-			|| (me->modifiers() & Qt::ShiftModifier) )
-		{
-			if( isSelected() )
-			{
-				m_action = MoveSelection;
-			}
-			else
-			{
-				gui->songEditor()->m_editor->selectAllTcos( false );
-				m_tco->addJournalCheckPoint();
-
-				// move or resize
-				m_tco->setJournalling( false );
-
-				setInitialPos( me->pos() );
-				setInitialOffsets();
-
-				SampleTCO * sTco = dynamic_cast<SampleTCO*>( m_tco );
-				if( me->x() < RESIZE_GRIP_WIDTH && sTco
-						&& !m_tco->getAutoResize() )
-				{
-					m_action = ResizeLeft;
-					setCursor( Qt::SizeHorCursor );
-				}
-				else if( m_tco->getAutoResize() || me->x() < width() - RESIZE_GRIP_WIDTH )
-				{
-					m_action = Move;
-					setCursor( Qt::SizeAllCursor );
-				}
-				else
-				{
-					m_action = Resize;
-					setCursor( Qt::SizeHorCursor );
-				}
-
-				if( m_action == Move )
-				{
-					s_textFloat->setTitle( tr( "Current position" ) );
-					s_textFloat->setText( QString( "%1:%2" ).
-						arg( m_tco->startPosition().getBar() + 1 ).
-						arg( m_tco->startPosition().getTicks() %
-								TimePos::ticksPerBar() ) );
-				}
-				else if( m_action == Resize || m_action == ResizeLeft )
-				{
-					s_textFloat->setTitle( tr( "Current length" ) );
-					s_textFloat->setText( tr( "%1:%2 (%3:%4 to %5:%6)" ).
-							arg( m_tco->length().getBar() ).
-							arg( m_tco->length().getTicks() %
-									TimePos::ticksPerBar() ).
-							arg( m_tco->startPosition().getBar() + 1 ).
-							arg( m_tco->startPosition().getTicks() %
-									TimePos::ticksPerBar() ).
-							arg( m_tco->endPosition().getBar() + 1 ).
-							arg( m_tco->endPosition().getTicks() %
-									TimePos::ticksPerBar() ) );
-				}
-				// s_textFloat->reparent( this );
-				// setup text-float as if TCO was already moved/resized
-				s_textFloat->moveGlobal( this, QPoint( width() + 2, height() + 2) );
-				s_textFloat->show();
-			}
-
-			delete m_hint;
-			QString hint = m_action == Move || m_action == MoveSelection
-						? tr( "Press <%1> and drag to make a copy." )
-						: tr( "Press <%1> for free resizing." );
-			m_hint = TextFloat::displayMessage( tr( "Hint" ), hint.arg(UI_CTRL_KEY),
-					embed::getIconPixmap( "hint" ), 0 );
-		}
-	}
-	else if( me->button() == Qt::RightButton )
-	{
-		if( me->modifiers() & Qt::ControlModifier )
-		{
-			toggleMute( active );
-		}
-		else if( me->modifiers() & Qt::ShiftModifier && !fixedTCOs() )
-		{
-			remove( active );
-		}
-	}
-	else if( me->button() == Qt::MidButton )
-	{
-		if( me->modifiers() & Qt::ControlModifier )
-		{
-			toggleMute( active );
-		}
-		else if( !fixedTCOs() )
-		{
-			remove( active );
-		}
-	}
-}
-
-
-
-
-/*! \brief Handle a mouse movement (drag) on this trackContentObjectView.
- *
- *  Handles the various ways in which a trackContentObjectView can be
- *  used with a mouse drag.
- *
- *  * If in move mode, move ourselves in the track,
- *  * or if in move-selection mode, move the entire selection,
- *  * or if in resize mode, resize ourselves,
- *  * otherwise ???
- *
- * \param me The QMouseEvent to handle.
- * \todo what does the final else case do here?
- */
-void TrackContentObjectView::mouseMoveEvent( QMouseEvent * me )
-{
-	if( m_action == CopySelection || m_action == ToggleSelected )
-	{
-		if( mouseMovedDistance( me, 2 ) == true )
-		{
-			QVector<TrackContentObjectView *> tcoViews;
-			if( m_action == CopySelection )
-			{
-				// Collect all selected TCOs
-				QVector<selectableObject *> so =
-					m_trackView->trackContainerView()->selectedObjects();
-				for( auto it = so.begin(); it != so.end(); ++it )
-				{
-					TrackContentObjectView * tcov =
-						dynamic_cast<TrackContentObjectView *>( *it );
-					if( tcov != NULL )
-					{
-						tcoViews.push_back( tcov );
-					}
-				}
-			}
-			else
-			{
-				gui->songEditor()->m_editor->selectAllTcos( false );
-				tcoViews.push_back( this );
-			}
-			// Clear the action here because mouseReleaseEvent will not get
-			// triggered once we go into drag.
-			m_action = NoAction;
-
-			// Write the TCOs to the DataFile for copying
-			DataFile dataFile = createTCODataFiles( tcoViews );
-
-			// TODO -- thumbnail for all selected
-			QPixmap thumbnail = grab().scaled(
-				128, 128,
-				Qt::KeepAspectRatio,
-				Qt::SmoothTransformation );
-			new StringPairDrag( QString( "tco_%1" ).arg(
-								m_tco->getTrack()->type() ),
-								dataFile.toString(), thumbnail, this );
-		}
-	}
-
-	if( me->modifiers() & Qt::ControlModifier )
-	{
-		delete m_hint;
-		m_hint = NULL;
-	}
-
-	const float ppb = m_trackView->trackContainerView()->pixelsPerBar();
-	if( m_action == Move )
-	{
-		TimePos newPos = draggedTCOPos( me );
-
-		m_tco->movePosition(newPos);
-		newPos = m_tco->startPosition(); // Get the real position the TCO was dragged to for the label
-		m_trackView->getTrackContentWidget()->changePosition();
-		s_textFloat->setText( QString( "%1:%2" ).
-				arg( newPos.getBar() + 1 ).
-				arg( newPos.getTicks() %
-						TimePos::ticksPerBar() ) );
-		s_textFloat->moveGlobal( this, QPoint( width() + 2, height() + 2 ) );
-	}
-	else if( m_action == MoveSelection )
-	{
-		// 1: Find the position we want to move the grabbed TCO to
-		TimePos newPos = draggedTCOPos( me );
-
-		// 2: Handle moving the other selected TCOs the same distance
-		QVector<selectableObject *> so =
-			m_trackView->trackContainerView()->selectedObjects();
-		QVector<TrackContentObject *> tcos; // List of selected clips
-		int leftmost = 0; // Leftmost clip's offset from grabbed clip
-		// Populate tcos, find leftmost
-		for( QVector<selectableObject *>::iterator it = so.begin();
-							it != so.end(); ++it )
-		{
-			TrackContentObjectView * tcov =
-				dynamic_cast<TrackContentObjectView *>( *it );
-			if( tcov == NULL ) { continue; }
-			tcos.push_back( tcov->m_tco );
-			int index = std::distance( so.begin(), it );
-			leftmost = min (leftmost, m_initialOffsets[index].getTicks() );
-		}
-		// Make sure the leftmost clip doesn't get moved to a negative position
-		if ( newPos.getTicks() + leftmost < 0 ) { newPos = -leftmost; }
-
-		for( QVector<TrackContentObject *>::iterator it = tcos.begin();
-							it != tcos.end(); ++it )
-		{
-			int index = std::distance( tcos.begin(), it );
-			( *it )->movePosition( newPos + m_initialOffsets[index] );
-		}
-	}
-	else if( m_action == Resize || m_action == ResizeLeft )
-	{
-		// If the user is holding alt, or pressed ctrl after beginning the drag, don't quantize
-		const bool unquantized = (me->modifiers() & Qt::ControlModifier) || (me->modifiers() & Qt::AltModifier);
-		const float snapSize = gui->songEditor()->m_editor->getSnapSize();
-		// Length in ticks of one snap increment
-		const TimePos snapLength = TimePos( (int)(snapSize * TimePos::ticksPerBar()) );
-
-		if( m_action == Resize )
-		{
-			// The clip's new length
-			TimePos l = static_cast<int>( me->x() * TimePos::ticksPerBar() / ppb );
-
-			if ( unquantized )
-			{	// We want to preserve this adjusted offset,
-				// even if the user switches to snapping later
-				setInitialPos( m_initialMousePos );
-				// Don't resize to less than 1 tick
-				m_tco->changeLength( qMax<int>( 1, l ) );
-			}
-			else if ( me->modifiers() & Qt::ShiftModifier )
-			{	// If shift is held, quantize clip's end position
-				TimePos end = TimePos( m_initialTCOPos + l ).quantize( snapSize );
-				// The end position has to be after the clip's start
-				TimePos min = m_initialTCOPos.quantize( snapSize );
-				if ( min <= m_initialTCOPos ) min += snapLength;
-				m_tco->changeLength( qMax<int>(min - m_initialTCOPos, end - m_initialTCOPos) );
-			}
-			else
-			{	// Otherwise, resize in fixed increments
-				TimePos initialLength = m_initialTCOEnd - m_initialTCOPos;
-				TimePos offset = TimePos( l - initialLength ).quantize( snapSize );
-				// Don't resize to less than 1 tick
-				TimePos min = TimePos( initialLength % snapLength );
-				if (min < 1) min += snapLength;
-				m_tco->changeLength( qMax<int>( min, initialLength + offset) );
-			}
-		}
-		else
-		{
-			SampleTCO * sTco = dynamic_cast<SampleTCO*>( m_tco );
-			if( sTco )
-			{
-				const int x = mapToParent( me->pos() ).x() - m_initialMousePos.x();
-
-				TimePos t = qMax( 0, (int)
-									m_trackView->trackContainerView()->currentPosition() +
-									static_cast<int>( x * TimePos::ticksPerBar() / ppb ) );
-
-				if( unquantized )
-				{	// We want to preserve this adjusted offset,
-					// even if the user switches to snapping later
-					setInitialPos( m_initialMousePos );
-					//Don't resize to less than 1 tick
-					t = qMin<int>( m_initialTCOEnd - 1, t);
-				}
-				else if( me->modifiers() & Qt::ShiftModifier )
-				{	// If shift is held, quantize clip's start position
-					// Don't let the start position move past the end position
-					TimePos max = m_initialTCOEnd.quantize( snapSize );
-					if ( max >= m_initialTCOEnd ) max -= snapLength;
-					t = qMin<int>( max, t.quantize( snapSize ) );
-				}
-				else
-				{	// Otherwise, resize in fixed increments
-					// Don't resize to less than 1 tick
-					TimePos initialLength = m_initialTCOEnd - m_initialTCOPos;
-					TimePos minLength = TimePos( initialLength % snapLength );
-					if (minLength < 1) minLength += snapLength;
-					TimePos offset = TimePos(t - m_initialTCOPos).quantize( snapSize );
-					t = qMin<int>( m_initialTCOEnd - minLength, m_initialTCOPos + offset );
-				}
-
-				TimePos oldPos = m_tco->startPosition();
-				if( m_tco->length() + ( oldPos - t ) >= 1 )
-				{
-					m_tco->movePosition( t );
-					m_tco->changeLength( m_tco->length() + ( oldPos - t ) );
-					sTco->setStartTimeOffset( sTco->startTimeOffset() + ( oldPos - t ) );
-				}
-			}
-		}
-		s_textFloat->setText( tr( "%1:%2 (%3:%4 to %5:%6)" ).
-				arg( m_tco->length().getBar() ).
-				arg( m_tco->length().getTicks() %
-						TimePos::ticksPerBar() ).
-				arg( m_tco->startPosition().getBar() + 1 ).
-				arg( m_tco->startPosition().getTicks() %
-						TimePos::ticksPerBar() ).
-				arg( m_tco->endPosition().getBar() + 1 ).
-				arg( m_tco->endPosition().getTicks() %
-						TimePos::ticksPerBar() ) );
-		s_textFloat->moveGlobal( this, QPoint( width() + 2, height() + 2) );
-	}
-	else
-	{
-		SampleTCO * sTco = dynamic_cast<SampleTCO*>( m_tco );
-		if( ( me->x() > width() - RESIZE_GRIP_WIDTH && !me->buttons() && !m_tco->getAutoResize() )
-		||  ( me->x() < RESIZE_GRIP_WIDTH && !me->buttons() && sTco && !m_tco->getAutoResize() ) )
-		{
-			setCursor( Qt::SizeHorCursor );
-		}
-		else
-		{
-			leaveEvent( NULL );
-		}
-	}
-}
-
-
-
-
-/*! \brief Handle a mouse release on this trackContentObjectView.
- *
- *  If we're in move or resize mode, journal the change as appropriate.
- *  Then tidy up.
- *
- * \param me The QMouseEvent to handle.
- */
-void TrackContentObjectView::mouseReleaseEvent( QMouseEvent * me )
-{
-	// If the CopySelection was chosen as the action due to mouse movement,
-	// it will have been cleared.  At this point Toggle is the desired action.
-	// An active StringPairDrag will prevent this method from being called,
-	// so a real CopySelection would not have occurred.
-	if( m_action == CopySelection ||
-	    ( m_action == ToggleSelected && mouseMovedDistance( me, 2 ) == false ) )
-	{
-		setSelected( !isSelected() );
-	}
-
-	if( m_action == Move || m_action == Resize || m_action == ResizeLeft )
-	{
-		// TODO: Fix m_tco->setJournalling() consistency
-		m_tco->setJournalling( true );
-	}
-	m_action = NoAction;
-	delete m_hint;
-	m_hint = NULL;
-	s_textFloat->hide();
-	leaveEvent( NULL );
-	selectableObject::mouseReleaseEvent( me );
-}
-
-
-
-
-/*! \brief Set up the context menu for this trackContentObjectView.
- *
- *  Set up the various context menu events that can apply to a
- *  track content object view.
- *
- * \param cme The QContextMenuEvent to add the actions to.
- */
-void TrackContentObjectView::contextMenuEvent( QContextMenuEvent * cme )
-{
-	// Depending on whether we right-clicked a selection or an individual TCO we will have
-	// different labels for the actions.
-	bool individualTCO = getClickedTCOs().size() <= 1;
-
-	if( cme->modifiers() )
-	{
-		return;
-	}
-
-	QMenu contextMenu( this );
-
-	if( fixedTCOs() == false )
-	{
-		contextMenu.addAction(
-			embed::getIconPixmap( "cancel" ),
-			individualTCO
-				? tr("Delete (middle mousebutton)")
-				: tr("Delete selection (middle mousebutton)"),
-			[this](){ contextMenuAction( Remove ); } );
-
-		contextMenu.addSeparator();
-
-		contextMenu.addAction(
-			embed::getIconPixmap( "edit_cut" ),
-			individualTCO
-				? tr("Cut")
-				: tr("Cut selection"),
-			[this](){ contextMenuAction( Cut ); } );
-	}
-
-	contextMenu.addAction(
-		embed::getIconPixmap( "edit_copy" ),
-		individualTCO
-			? tr("Copy")
-			: tr("Copy selection"),
-		[this](){ contextMenuAction( Copy ); } );
-
-	contextMenu.addAction(
-		embed::getIconPixmap( "edit_paste" ),
-		tr( "Paste" ),
-		[this](){ contextMenuAction( Paste ); } );
-
-	contextMenu.addSeparator();
-
-	contextMenu.addAction(
-		embed::getIconPixmap( "muted" ),
-		(individualTCO
-			? tr("Mute/unmute (<%1> + middle click)")
-			: tr("Mute/unmute selection (<%1> + middle click)")).arg(UI_CTRL_KEY),
-		[this](){ contextMenuAction( Mute ); } );
-
-	contextMenu.addSeparator();
-
-	contextMenu.addAction( embed::getIconPixmap( "colorize" ),
-			tr( "Set clip color" ), this, SLOT( changeClipColor() ) );
-	contextMenu.addAction( embed::getIconPixmap( "colorize" ),
-			tr( "Use track color" ), this, SLOT( useTrackColor() ) );
-
-	constructContextMenu( &contextMenu );
-
-	contextMenu.exec( QCursor::pos() );
-}
-
-// This method processes the actions from the context menu of the TCO View.
-void TrackContentObjectView::contextMenuAction( ContextMenuAction action )
-{
-	QVector<TrackContentObjectView *> active = getClickedTCOs();
-	// active will be later used for the remove, copy, cut or toggleMute methods
-
-	switch( action )
-	{
-		case Remove:
-			remove( active );
-			break;
-		case Cut:
-			cut( active );
-			break;
-		case Copy:
-			copy( active );
-			break;
-		case Paste:
-			paste();
-			break;
-		case Mute:
-			toggleMute( active );
-			break;
-	}
-}
-
-QVector<TrackContentObjectView *> TrackContentObjectView::getClickedTCOs()
-{
-	// Get a list of selected selectableObjects
-	QVector<selectableObject *> sos = gui->songEditor()->m_editor->selectedObjects();
-
-	// Convert to a list of selected TCOVs
-	QVector<TrackContentObjectView *> selection;
-	selection.reserve( sos.size() );
-	for( auto so: sos )
-	{
-		TrackContentObjectView *tcov = dynamic_cast<TrackContentObjectView *> ( so );
-		if( tcov != nullptr )
-		{
-			selection.append( tcov );
-		}
-	}
-
-	// If we clicked part of the selection, affect all selected clips. Otherwise affect the clip we clicked
-	return selection.contains(this)
-		? selection
-		: QVector<TrackContentObjectView *>( 1, this );
-}
-
-void TrackContentObjectView::remove( QVector<TrackContentObjectView *> tcovs )
-{
-	for( auto tcov: tcovs )
-	{
-		// No need to check if it's nullptr because we check when building the QVector
-		tcov->remove();
-	}
-}
-
-void TrackContentObjectView::copy( QVector<TrackContentObjectView *> tcovs )
-{
-	// For copyStringPair()
-	using namespace Clipboard;
-
-	// Write the TCOs to a DataFile for copying
-	DataFile dataFile = createTCODataFiles( tcovs );
-
-	// Copy the TCO type as a key and the TCO data file to the clipboard
-	copyStringPair( QString( "tco_%1" ).arg( m_tco->getTrack()->type() ),
-		dataFile.toString() );
-}
-
-void TrackContentObjectView::cut( QVector<TrackContentObjectView *> tcovs )
-{
-	// Copy the selected TCOs
-	copy( tcovs );
-
-	// Now that the TCOs are copied we can delete them, since we are cutting
-	remove( tcovs );
-}
-
-void TrackContentObjectView::paste()
-{
-	// For getMimeData()
-	using namespace Clipboard;
-
-	// If possible, paste the selection on the TimePos of the selected Track and remove it
-	TimePos tcoPos = TimePos( m_tco->startPosition() );
-
-	TrackContentWidget *tcw = getTrackView()->getTrackContentWidget();
-
-	if( tcw->pasteSelection( tcoPos, getMimeData() ) )
-	{
-		// If we succeed on the paste we delete the TCO we pasted on
-		remove();
-	}
-}
-
-void TrackContentObjectView::toggleMute( QVector<TrackContentObjectView *> tcovs )
-{
-	for( auto tcov: tcovs )
-	{
-		// No need to check for nullptr because we check while building the tcovs QVector
-		tcov->getTrackContentObject()->toggleMute();
-	}
-}
-
-
-
-
-/*! \brief How many pixels a bar takes for this trackContentObjectView.
- *
- * \return the number of pixels per bar.
- */
-float TrackContentObjectView::pixelsPerBar()
-{
-	return m_trackView->trackContainerView()->pixelsPerBar();
-}
-
-
-/*! \brief Save the offsets between all selected tracks and a clicked track */
-void TrackContentObjectView::setInitialOffsets()
-{
-	QVector<selectableObject *> so = m_trackView->trackContainerView()->selectedObjects();
-	QVector<TimePos> offsets;
-	for( QVector<selectableObject *>::iterator it = so.begin();
-						it != so.end(); ++it )
-	{
-		TrackContentObjectView * tcov =
-			dynamic_cast<TrackContentObjectView *>( *it );
-		if( tcov == NULL )
-		{
-			continue;
-		}
-		offsets.push_back( tcov->m_tco->startPosition() - m_initialTCOPos );
-	}
-
-	m_initialOffsets = offsets;
-}
-
-
-
-
-/*! \brief Detect whether the mouse moved more than n pixels on screen.
- *
- * \param _me The QMouseEvent.
- * \param distance The threshold distance that the mouse has moved to return true.
- */
-bool TrackContentObjectView::mouseMovedDistance( QMouseEvent * me, int distance )
-{
-	QPoint dPos = mapToGlobal( me->pos() ) - m_initialMouseGlobalPos;
-	const int pixelsMoved = dPos.manhattanLength();
-	return ( pixelsMoved > distance || pixelsMoved < -distance );
-}
-
-
-
-/*! \brief Calculate the new position of a dragged TCO from a mouse event
- *
- *
- * \param me The QMouseEvent
- */
-TimePos TrackContentObjectView::draggedTCOPos( QMouseEvent * me )
-{
-	//Pixels per bar
-	const float ppb = m_trackView->trackContainerView()->pixelsPerBar();
-	// The pixel distance that the mouse has moved
-	const int mouseOff = mapToGlobal(me->pos()).x() - m_initialMouseGlobalPos.x();
-	TimePos newPos = m_initialTCOPos + mouseOff * TimePos::ticksPerBar() / ppb;
-	TimePos offset = newPos - m_initialTCOPos;
-	// If the user is holding alt, or pressed ctrl after beginning the drag, don't quantize
-	if (    me->button() != Qt::NoButton
-		|| (me->modifiers() & Qt::ControlModifier)
-		|| (me->modifiers() & Qt::AltModifier)    )
-	{
-		// We want to preserve this adjusted offset,
-		// even if the user switches to snapping
-		setInitialPos( m_initialMousePos );
-	}
-	else if ( me->modifiers() & Qt::ShiftModifier )
-	{	// If shift is held, quantize position (Default in 1.2.0 and earlier)
-		// or end position, whichever is closest to the actual position
-		TimePos startQ = newPos.quantize( gui->songEditor()->m_editor->getSnapSize() );
-		// Find start position that gives snapped clip end position
-		TimePos endQ = ( newPos + m_tco->length() );
-		endQ = endQ.quantize( gui->songEditor()->m_editor->getSnapSize() );
-		endQ = endQ - m_tco->length();
-		// Select the position closest to actual position
-		if ( abs(newPos - startQ) < abs(newPos - endQ) ) newPos = startQ;
-		else newPos = endQ;
-	}
-	else
-	{	// Otherwise, quantize moved distance (preserves user offsets)
-		newPos = m_initialTCOPos + offset.quantize( gui->songEditor()->m_editor->getSnapSize() );
-	}
-	return newPos;
-}
-
-
-// Return the color that the TCO's background should be
-QColor TrackContentObjectView::getColorForDisplay( QColor defaultColor )
-{
-	// Get the pure TCO color
-	auto tcoColor = m_tco->hasColor()
-					? m_tco->usesCustomClipColor()
-						? m_tco->color()
-						: m_tco->getTrack()->color()
-					: defaultColor;
-
-	// Set variables
-	QColor c, mutedCustomColor;
-	bool muted = m_tco->getTrack()->isMuted() || m_tco->isMuted();
-	mutedCustomColor = tcoColor;
-	mutedCustomColor.setHsv( mutedCustomColor.hsvHue(), mutedCustomColor.hsvSaturation() / 4, mutedCustomColor.value() );
-
-	// Change the pure color by state: selected, muted, colored, normal
-	if( isSelected() )
-	{
-		c = m_tco->hasColor()
-			? ( muted
-				? mutedCustomColor.darker( 350 )
-				: tcoColor.darker( 150 ) )
-			: selectedColor();
-	}
-	else
-	{
-		if( muted )
-		{
-			c = m_tco->hasColor()
-				? mutedCustomColor.darker( 250 )
-				: mutedBackgroundColor();
-		}
-		else
-		{
-			c = tcoColor;
-		}
-	}
-	
-	// Return color to caller
-	return c;
-}
-
-
-
-
-
-// ===========================================================================
-// trackContentWidget
-// ===========================================================================
-/*! \brief Create a new trackContentWidget
- *
- *  Creates a new track content widget for the given track.
- *  The content widget comprises the 'grip bar' and the 'tools' button
- *  for the track's context menu.
- *
- * \param parent The parent track.
- */
-TrackContentWidget::TrackContentWidget( TrackView * parent ) :
-	QWidget( parent ),
-	m_trackView( parent ),
-	m_darkerColor( Qt::SolidPattern ),
-	m_lighterColor( Qt::SolidPattern ),
-	m_gridColor( Qt::SolidPattern ),
-	m_embossColor( Qt::SolidPattern )
-{
-	setAcceptDrops( true );
-
-	connect( parent->trackContainerView(),
-			SIGNAL( positionChanged( const TimePos & ) ),
-			this, SLOT( changePosition( const TimePos & ) ) );
-
-	setStyle( QApplication::style() );
-
-	updateBackground();
-}
-
-
-
-
-/*! \brief Destroy this trackContentWidget
- *
- *  Destroys the trackContentWidget.
- */
-TrackContentWidget::~TrackContentWidget()
-{
-}
-
-
-
-
-void TrackContentWidget::updateBackground()
-{
-	const TrackContainerView * tcv = m_trackView->trackContainerView();
-
-	// Assume even-pixels-per-bar. Makes sense, should be like this anyways
-	int ppb = static_cast<int>( tcv->pixelsPerBar() );
-
-	int w = ppb * BARS_PER_GROUP;
-	int h = height();
-	m_background = QPixmap( w * 2, height() );
-	QPainter pmp( &m_background );
-
-	pmp.fillRect( 0, 0, w, h, darkerColor() );
-	pmp.fillRect( w, 0, w , h, lighterColor() );
-
-	// draw lines
-	// vertical lines
-	pmp.setPen( QPen( gridColor(), 1 ) );
-	for( float x = 0; x < w * 2; x += ppb )
-	{
-		pmp.drawLine( QLineF( x, 0.0, x, h ) );
-	}
-
-	pmp.setPen( QPen( embossColor(), 1 ) );
-	for( float x = 1.0; x < w * 2; x += ppb )
-	{
-		pmp.drawLine( QLineF( x, 0.0, x, h ) );
-	}
-
-	// horizontal line
-	pmp.setPen( QPen( gridColor(), 1 ) );
-	pmp.drawLine( 0, h-1, w*2, h-1 );
-
-	pmp.end();
-
-	// Force redraw
-	update();
-}
-
-
-
-
-/*! \brief Adds a trackContentObjectView to this widget.
- *
- *  Adds a(nother) trackContentObjectView to our list of views.  We also
- *  check that our position is up-to-date.
- *
- * \param tcov The trackContentObjectView to add.
- */
-void TrackContentWidget::addTCOView( TrackContentObjectView * tcov )
-{
-	TrackContentObject * tco = tcov->getTrackContentObject();
-
-	m_tcoViews.push_back( tcov );
-
-	tco->saveJournallingState( false );
-	changePosition();
-	tco->restoreJournallingState();
-}
-
-
-
-
-/*! \brief Removes the given trackContentObjectView to this widget.
- *
- *  Removes the given trackContentObjectView from our list of views.
- *
- * \param tcov The trackContentObjectView to add.
- */
-void TrackContentWidget::removeTCOView( TrackContentObjectView * tcov )
-{
-	tcoViewVector::iterator it = std::find( m_tcoViews.begin(),
-						m_tcoViews.end(),
-						tcov );
-	if( it != m_tcoViews.end() )
-	{
-		m_tcoViews.erase( it );
-		Engine::getSong()->setModified();
-	}
-}
-
-
-
-
-/*! \brief Update ourselves by updating all the tCOViews attached.
- *
- */
-void TrackContentWidget::update()
-{
-	for( tcoViewVector::iterator it = m_tcoViews.begin();
-				it != m_tcoViews.end(); ++it )
-	{
-		( *it )->setFixedHeight( height() - 1 );
-		( *it )->update();
-	}
-	QWidget::update();
-}
-
-
-
-
-// resposible for moving track-content-widgets to appropriate position after
-// change of visible viewport
-/*! \brief Move the trackContentWidget to a new place in time
- *
- * \param newPos The MIDI time to move to.
- */
-void TrackContentWidget::changePosition( const TimePos & newPos )
-{
-	if( m_trackView->trackContainerView() == gui->getBBEditor()->trackContainerView() )
-	{
-		const int curBB = Engine::getBBTrackContainer()->currentBB();
-		setUpdatesEnabled( false );
-
-		// first show TCO for current BB...
-		for( tcoViewVector::iterator it = m_tcoViews.begin();
-						it != m_tcoViews.end(); ++it )
-		{
-		if( ( *it )->getTrackContentObject()->
-						startPosition().getBar() == curBB )
-			{
-				( *it )->move( 0, ( *it )->y() );
-				( *it )->raise();
-				( *it )->show();
-			}
-			else
-			{
-				( *it )->lower();
-			}
-		}
-		// ...then hide others to avoid flickering
-		for( tcoViewVector::iterator it = m_tcoViews.begin();
-					it != m_tcoViews.end(); ++it )
-		{
-			if( ( *it )->getTrackContentObject()->
-						startPosition().getBar() != curBB )
-			{
-				( *it )->hide();
-			}
-		}
-		setUpdatesEnabled( true );
-		return;
-	}
-
-	TimePos pos = newPos;
-	if( pos < 0 )
-	{
-		pos = m_trackView->trackContainerView()->currentPosition();
-	}
-
-	const int begin = pos;
-	const int end = endPosition( pos );
-	const float ppb = m_trackView->trackContainerView()->pixelsPerBar();
-
-	setUpdatesEnabled( false );
-	for( tcoViewVector::iterator it = m_tcoViews.begin();
-						it != m_tcoViews.end(); ++it )
-	{
-		TrackContentObjectView * tcov = *it;
-		TrackContentObject * tco = tcov->getTrackContentObject();
-
-		tco->changeLength( tco->length() );
-
-		const int ts = tco->startPosition();
-		const int te = tco->endPosition()-3;
-		if( ( ts >= begin && ts <= end ) ||
-			( te >= begin && te <= end ) ||
-			( ts <= begin && te >= end ) )
-		{
-			tcov->move( static_cast<int>( ( ts - begin ) * ppb /
-						TimePos::ticksPerBar() ),
-								tcov->y() );
-			if( !tcov->isVisible() )
-			{
-				tcov->show();
-			}
-		}
-		else
-		{
-			tcov->move( -tcov->width()-10, tcov->y() );
-		}
-	}
-	setUpdatesEnabled( true );
-
-	// redraw background
-//	update();
-}
-
-
-
-
-/*! \brief Return the position of the trackContentWidget in bars.
- *
- * \param mouseX the mouse's current X position in pixels.
- */
-TimePos TrackContentWidget::getPosition( int mouseX )
-{
-	TrackContainerView * tv = m_trackView->trackContainerView();
-	return TimePos( tv->currentPosition() +
-					 mouseX *
-					 TimePos::ticksPerBar() /
-					 static_cast<int>( tv->pixelsPerBar() ) );
-}
-
-
-
-
-/*! \brief Respond to a drag enter event on the trackContentWidget
- *
- * \param dee the Drag Enter Event to respond to
- */
-void TrackContentWidget::dragEnterEvent( QDragEnterEvent * dee )
-{
-	TimePos tcoPos = getPosition( dee->pos().x() );
-  if( canPasteSelection( tcoPos, dee ) == false )
-	{
-		dee->ignore();
-	}
-	else
-	{
-		StringPairDrag::processDragEnterEvent( dee, "tco_" +
-						QString::number( getTrack()->type() ) );
-	}
-}
-
-
-
-
-/*! \brief Returns whether a selection of TCOs can be pasted into this
- *
- * \param tcoPos the position of the TCO slot being pasted on
- * \param de the DropEvent generated
- */
-bool TrackContentWidget::canPasteSelection( TimePos tcoPos, const QDropEvent* de )
-{
-	const QMimeData * mimeData = de->mimeData();
-
-	// If the source of the DropEvent is the current instance of LMMS we don't allow pasting in the same bar
-	// if it's another instance of LMMS we allow it
-	return de->source()
-		? canPasteSelection( tcoPos, mimeData )
-		: canPasteSelection( tcoPos, mimeData, true );
-}
-
-// Overloaded method to make it possible to call this method without a Drag&Drop event
-bool TrackContentWidget::canPasteSelection( TimePos tcoPos, const QMimeData* md , bool allowSameBar )
-{
-	// For decodeKey() and decodeValue()
-	using namespace Clipboard;
-
-	Track * t = getTrack();
-	QString type = decodeKey( md );
-	QString value = decodeValue( md );
-
-	// We can only paste into tracks of the same type
-	if( type != ( "tco_" + QString::number( t->type() ) ) ||
-		m_trackView->trackContainerView()->fixedTCOs() == true )
-	{
-		return false;
-	}
-
-	// value contains XML needed to reconstruct TCOs and place them
-	DataFile dataFile( value.toUtf8() );
-
-	// Extract the metadata and which TCO was grabbed
-	QDomElement metadata = dataFile.content().firstChildElement( "copyMetadata" );
-	QDomAttr tcoPosAttr = metadata.attributeNode( "grabbedTCOPos" );
-	TimePos grabbedTCOPos = tcoPosAttr.value().toInt();
-	TimePos grabbedTCOBar = TimePos( grabbedTCOPos.getBar(), 0 );
-
-	// Extract the track index that was originally clicked
-	QDomAttr tiAttr = metadata.attributeNode( "initialTrackIndex" );
-	const int initialTrackIndex = tiAttr.value().toInt();
-
-	// Get the current track's index
-	const TrackContainer::TrackList tracks = t->trackContainer()->tracks();
-	const int currentTrackIndex = tracks.indexOf( t );
-
-	// Don't paste if we're on the same bar and allowSameBar is false
-	auto sourceTrackContainerId = metadata.attributeNode( "trackContainerId" ).value().toUInt();
-	if( !allowSameBar && sourceTrackContainerId == t->trackContainer()->id() &&
-			tcoPos == grabbedTCOBar && currentTrackIndex == initialTrackIndex )
-	{
-		return false;
-	}
-
-	// Extract the tco data
-	QDomElement tcoParent = dataFile.content().firstChildElement( "tcos" );
-	QDomNodeList tcoNodes = tcoParent.childNodes();
-
-	// Determine if all the TCOs will land on a valid track
-	for( int i = 0; i < tcoNodes.length(); i++ )
-	{
-		QDomElement tcoElement = tcoNodes.item( i ).toElement();
-		int trackIndex = tcoElement.attributeNode( "trackIndex" ).value().toInt();
-		int finalTrackIndex = trackIndex + currentTrackIndex - initialTrackIndex;
-
-		// Track must be in TrackContainer's tracks
-		if( finalTrackIndex < 0 || finalTrackIndex >= tracks.size() )
-		{
-			return false;
-		}
-
-		// Track must be of the same type
-		auto startTrackType = tcoElement.attributeNode("trackType").value().toInt();
-		Track * endTrack = tracks.at( finalTrackIndex );
-		if( startTrackType != endTrack->type() )
-		{
-			return false;
-		}
-	}
-
-	return true;
-}
-
-/*! \brief Pastes a selection of TCOs onto the track
- *
- * \param tcoPos the position of the TCO slot being pasted on
- * \param de the DropEvent generated
- */
-bool TrackContentWidget::pasteSelection( TimePos tcoPos, QDropEvent * de )
-{
-	const QMimeData * mimeData = de->mimeData();
-
-	if( canPasteSelection( tcoPos, de ) == false )
-	{
-		return false;
-	}
-
-	// We set skipSafetyCheck to true because we already called canPasteSelection
-	return pasteSelection( tcoPos, mimeData, true );
-}
-
-// Overloaded method so we can call it without a Drag&Drop event
-bool TrackContentWidget::pasteSelection( TimePos tcoPos, const QMimeData * md, bool skipSafetyCheck )
-{
-	// For decodeKey() and decodeValue()
-	using namespace Clipboard;
-
-	// When canPasteSelection was already called before, skipSafetyCheck will skip this
-	if( !skipSafetyCheck && canPasteSelection( tcoPos, md ) == false )
-	{
-		return false;
-	}
-
-	QString type = decodeKey( md );
-	QString value = decodeValue( md );
-
-	getTrack()->addJournalCheckPoint();
-
-	// value contains XML needed to reconstruct TCOs and place them
-	DataFile dataFile( value.toUtf8() );
-
-	// Extract the tco data
-	QDomElement tcoParent = dataFile.content().firstChildElement( "tcos" );
-	QDomNodeList tcoNodes = tcoParent.childNodes();
-
-	// Extract the track index that was originally clicked
-	QDomElement metadata = dataFile.content().firstChildElement( "copyMetadata" );
-	QDomAttr tiAttr = metadata.attributeNode( "initialTrackIndex" );
-	int initialTrackIndex = tiAttr.value().toInt();
-	QDomAttr tcoPosAttr = metadata.attributeNode( "grabbedTCOPos" );
-	TimePos grabbedTCOPos = tcoPosAttr.value().toInt();
-
-	// Snap the mouse position to the beginning of the dropped bar, in ticks
-	const TrackContainer::TrackList tracks = getTrack()->trackContainer()->tracks();
-	const int currentTrackIndex = tracks.indexOf( getTrack() );
-
-	bool wasSelection = m_trackView->trackContainerView()->rubberBand()->selectedObjects().count();
-
-	// Unselect the old group
-		const QVector<selectableObject *> so =
-			m_trackView->trackContainerView()->selectedObjects();
-		for( QVector<selectableObject *>::const_iterator it = so.begin();
-		    	it != so.end(); ++it )
-		{
-			( *it )->setSelected( false );
-		}
-
-
-	// TODO -- Need to draw the hovericon either way, or ghost the TCOs
-	// onto their final position.
-
-	float snapSize = gui->songEditor()->m_editor->getSnapSize();
-	// All patterns should be offset the same amount as the grabbed pattern
-	TimePos offset = TimePos(tcoPos - grabbedTCOPos);
-	// Users expect clips to "fall" backwards, so bias the offset
-	offset = offset - TimePos::ticksPerBar() * snapSize / 2;
-	// The offset is quantized (rather than the positions) to preserve fine adjustments
-	offset = offset.quantize(snapSize);
-
-	// Get the leftmost TCO and fix the offset if it reaches below bar 0
-	TimePos leftmostPos = grabbedTCOPos;
-	for(int i = 0; i < tcoNodes.length(); ++i)
-	{
-		QDomElement outerTCOElement = tcoNodes.item(i).toElement();
-		QDomElement tcoElement = outerTCOElement.firstChildElement();
-
-		TimePos pos = tcoElement.attributeNode("pos").value().toInt();
-
-		if(pos < leftmostPos) { leftmostPos = pos; }
-	}
-	// Fix offset if it sets the left most TCO to a negative position
-	offset = std::max(offset.getTicks(), -leftmostPos.getTicks());
-
-	for( int i = 0; i<tcoNodes.length(); i++ )
-	{
-		QDomElement outerTCOElement = tcoNodes.item( i ).toElement();
-		QDomElement tcoElement = outerTCOElement.firstChildElement();
-
-		int trackIndex = outerTCOElement.attributeNode( "trackIndex" ).value().toInt();
-		int finalTrackIndex = trackIndex + ( currentTrackIndex - initialTrackIndex );
-		Track * t = tracks.at( finalTrackIndex );
-
-		// The new position is the old position plus the offset.
-		TimePos pos = tcoElement.attributeNode( "pos" ).value().toInt() + offset;
-		// If we land on ourselves, offset by one snap
-		TimePos shift = TimePos::ticksPerBar() * gui->songEditor()->m_editor->getSnapSize();
-		if (offset == 0) { pos += shift; }
-
-		TrackContentObject * tco = t->createTCO( pos );
-		tco->restoreState( tcoElement );
-		tco->movePosition(pos); // Because we restored the state, we need to move the TCO again.
-		if( wasSelection )
-		{
-			tco->selectViewOnCreate( true );
-		}
-	}
-
-	AutomationPattern::resolveAllIDs();
-
-	return true;
-}
-
-
-/*! \brief Respond to a drop event on the trackContentWidget
- *
- * \param de the Drop Event to respond to
- */
-void TrackContentWidget::dropEvent( QDropEvent * de )
-{
-	TimePos tcoPos = TimePos( getPosition( de->pos().x() ) );
-	if( pasteSelection( tcoPos, de ) == true )
-	{
-		de->accept();
-	}
-}
-
-
-
-
-/*! \brief Respond to a mouse press on the trackContentWidget
- *
- * \param me the mouse press event to respond to
- */
-void TrackContentWidget::mousePressEvent( QMouseEvent * me )
-{
-	if( m_trackView->trackContainerView()->allowRubberband() == true )
-	{
-		QWidget::mousePressEvent( me );
-	}
-	else if( me->modifiers() & Qt::ShiftModifier )
-	{
-		QWidget::mousePressEvent( me );
-	}
-	else if( me->button() == Qt::LeftButton &&
-			!m_trackView->trackContainerView()->fixedTCOs() )
-	{
-		QVector<selectableObject*> so =  m_trackView->trackContainerView()->rubberBand()->selectedObjects();
-		for( int i = 0; i < so.count(); ++i )
-		{
-			so.at( i )->setSelected( false);
-		}
-		getTrack()->addJournalCheckPoint();
-		const TimePos pos = getPosition( me->x() ).getBar() *
-						TimePos::ticksPerBar();
-		getTrack()->createTCO(pos);
-	}
-}
-
-
-
-
-/*! \brief Repaint the trackContentWidget on command
- *
- * \param pe the Paint Event to respond to
- */
-void TrackContentWidget::paintEvent( QPaintEvent * pe )
-{
-	// Assume even-pixels-per-bar. Makes sense, should be like this anyways
-	const TrackContainerView * tcv = m_trackView->trackContainerView();
-	int ppb = static_cast<int>( tcv->pixelsPerBar() );
-	QPainter p( this );
-	// Don't draw background on BB-Editor
-	if( m_trackView->trackContainerView() != gui->getBBEditor()->trackContainerView() )
-	{
-		p.drawTiledPixmap( rect(), m_background, QPoint(
-				tcv->currentPosition().getBar() * ppb, 0 ) );
-	}
-}
-
-
-
-
-/*! \brief Updates the background tile pixmap on size changes.
- *
- * \param resizeEvent the resize event to pass to base class
- */
-void TrackContentWidget::resizeEvent( QResizeEvent * resizeEvent )
-{
-	// Update backgroud
-	updateBackground();
-	// Force redraw
-	QWidget::resizeEvent( resizeEvent );
-}
-
-
-
-
-/*! \brief Return the track shown by the trackContentWidget
- *
- */
-Track * TrackContentWidget::getTrack()
-{
-	return m_trackView->getTrack();
-}
-
-
-
-
-/*! \brief Return the end position of the trackContentWidget in Bars.
- *
- * \param posStart the starting position of the Widget (from getPosition())
- */
-TimePos TrackContentWidget::endPosition( const TimePos & posStart )
-{
-	const float ppb = m_trackView->trackContainerView()->pixelsPerBar();
-	const int w = width();
-	return posStart + static_cast<int>( w * TimePos::ticksPerBar() / ppb );
-}
-
-void TrackContentWidget::contextMenuEvent( QContextMenuEvent * cme )
-{
-	// For hasFormat(), MimeType enum class and getMimeData()
-	using namespace Clipboard;
-
-	if( cme->modifiers() )
-	{
-		return;
-	}
-
-	// If we don't have TCO data in the clipboard there's no need to create this menu
-	// since "paste" is the only action at the moment.
-	if( ! hasFormat( MimeType::StringPair )  )
-	{
-		return;
-	}
-
-	QMenu contextMenu( this );
-	QAction *pasteA = contextMenu.addAction( embed::getIconPixmap( "edit_paste" ),
-					tr( "Paste" ), [this, cme](){ contextMenuAction( cme, Paste ); } );
-	// If we can't paste in the current TCW for some reason, disable the action so the user knows
-	pasteA->setEnabled( canPasteSelection( getPosition( cme->x() ), getMimeData() ) ? true : false );
-
-	contextMenu.exec( QCursor::pos() );
-}
-
-void TrackContentWidget::contextMenuAction( QContextMenuEvent * cme, ContextMenuAction action )
-{
-	// For getMimeData()
-	using namespace Clipboard;
-
-	switch( action )
-	{
-		case Paste:
-		// Paste the selection on the TimePos of the context menu event
-		TimePos tcoPos = getPosition( cme->x() );
-
-		pasteSelection( tcoPos, getMimeData() );
-		break;
-	}
-}
-
-
-
-// qproperty access methods
-//! \brief CSS theming qproperty access method
-QBrush TrackContentWidget::darkerColor() const
-{ return m_darkerColor; }
-
-//! \brief CSS theming qproperty access method
-QBrush TrackContentWidget::lighterColor() const
-{ return m_lighterColor; }
-
-//! \brief CSS theming qproperty access method
-QBrush TrackContentWidget::gridColor() const
-{ return m_gridColor; }
-
-//! \brief CSS theming qproperty access method
-QBrush TrackContentWidget::embossColor() const
-{ return m_embossColor; }
-
-//! \brief CSS theming qproperty access method
-void TrackContentWidget::setDarkerColor( const QBrush & c )
-{ m_darkerColor = c; }
-
-//! \brief CSS theming qproperty access method
-void TrackContentWidget::setLighterColor( const QBrush & c )
-{ m_lighterColor = c; }
-
-//! \brief CSS theming qproperty access method
-void TrackContentWidget::setGridColor( const QBrush & c )
-{ m_gridColor = c; }
-
-//! \brief CSS theming qproperty access method
-void TrackContentWidget::setEmbossColor( const QBrush & c )
-{ m_embossColor = c; }
-
-
-// ===========================================================================
-// trackOperationsWidget
-// ===========================================================================
-
-/*! \brief Create a new trackOperationsWidget
- *
- * The trackOperationsWidget is the grip and the mute button of a track.
- *
- * \param parent the trackView to contain this widget
- */
-TrackOperationsWidget::TrackOperationsWidget( TrackView * parent ) :
-	QWidget( parent ),             /*!< The parent widget */
-	m_trackView( parent )          /*!< The parent track view */
-{
-	ToolTip::add( this, tr( "Press <%1> while clicking on move-grip "
-				"to begin a new drag'n'drop action." ).arg(UI_CTRL_KEY) );
-
-	QMenu * toMenu = new QMenu( this );
-	toMenu->setFont( pointSize<9>( toMenu->font() ) );
-	connect( toMenu, SIGNAL( aboutToShow() ), this, SLOT( updateMenu() ) );
-
-
-	setObjectName( "automationEnabled" );
-
-
-	m_trackOps = new QPushButton( this );
-	m_trackOps->move( 12, 1 );
-	m_trackOps->setFocusPolicy( Qt::NoFocus );
-	m_trackOps->setMenu( toMenu );
-	ToolTip::add( m_trackOps, tr( "Actions" ) );
-
-
-	m_muteBtn = new PixmapButton( this, tr( "Mute" ) );
-	m_muteBtn->setActiveGraphic( embed::getIconPixmap( "led_off" ) );
-	m_muteBtn->setInactiveGraphic( embed::getIconPixmap( "led_green" ) );
-	m_muteBtn->setCheckable( true );
-
-	m_soloBtn = new PixmapButton( this, tr( "Solo" ) );
-	m_soloBtn->setActiveGraphic( embed::getIconPixmap( "led_red" ) );
-	m_soloBtn->setInactiveGraphic( embed::getIconPixmap( "led_off" ) );
-	m_soloBtn->setCheckable( true );
-
-	if( ConfigManager::inst()->value( "ui",
-					  "compacttrackbuttons" ).toInt() )
-	{
-		m_muteBtn->move( 46, 0 );
-		m_soloBtn->move( 46, 16 );
-	}
-	else
-	{
-		m_muteBtn->move( 46, 8 );
-		m_soloBtn->move( 62, 8 );
-	}
-
-	m_muteBtn->show();
-	ToolTip::add( m_muteBtn, tr( "Mute" ) );
-
-	m_soloBtn->show();
-	ToolTip::add( m_soloBtn, tr( "Solo" ) );
-
-	connect( this, SIGNAL( trackRemovalScheduled( TrackView * ) ),
-			m_trackView->trackContainerView(),
-				SLOT( deleteTrackView( TrackView * ) ),
-							Qt::QueuedConnection );
-	
-	connect( m_trackView->getTrack()->getMutedModel(), SIGNAL( dataChanged() ),
-			this, SLOT( update() ) );
-	
-}
-
-
-
-
-/*! \brief Destroy an existing trackOperationsWidget
- *
- */
-TrackOperationsWidget::~TrackOperationsWidget()
-{
-}
-
-
-
-
-/*! \brief Respond to trackOperationsWidget mouse events
- *
- *  If it's the left mouse button, and Ctrl is held down, and we're
- *  not a Beat+Bassline Editor track, then start a new drag event to
- *  copy this track.
- *
- *  Otherwise, ignore all other events.
- *
- *  \param me The mouse event to respond to.
- */
-void TrackOperationsWidget::mousePressEvent( QMouseEvent * me )
-{
-	if( me->button() == Qt::LeftButton &&
-		me->modifiers() & Qt::ControlModifier &&
-			m_trackView->getTrack()->type() != Track::BBTrack )
-	{
-		DataFile dataFile( DataFile::DragNDropData );
-		m_trackView->getTrack()->saveState( dataFile, dataFile.content() );
-		new StringPairDrag( QString( "track_%1" ).arg(
-					m_trackView->getTrack()->type() ),
-			dataFile.toString(), m_trackView->getTrackSettingsWidget()->grab(),
-									this );
-	}
-	else if( me->button() == Qt::LeftButton )
-	{
-		// track-widget (parent-widget) initiates track-move
-		me->ignore();
-	}
-}
-
-
-
-
-/*! \brief Repaint the trackOperationsWidget
- *
- *  If we're not moving, and in the Beat+Bassline Editor, then turn
- *  automation on or off depending on its previous state and show
- *  ourselves.
- *
- *  Otherwise, hide ourselves.
- *
- *  \todo Flesh this out a bit - is it correct?
- *  \param pe The paint event to respond to
- */
-void TrackOperationsWidget::paintEvent( QPaintEvent * pe )
-{
-	QPainter p( this );
-	
-	p.fillRect( rect(), palette().brush(QPalette::Background) );
-	
-	if( m_trackView->getTrack()->useColor() && ! m_trackView->getTrack()->getMutedModel()->value() ) 
-	{
-		QRect coloredRect( 0, 0, 10, m_trackView->getTrack()->getHeight() );
-		
-		p.fillRect( coloredRect, m_trackView->getTrack()->color() );
-	}
-
-	if( m_trackView->isMovingTrack() == false )
-	{
-		p.drawPixmap( 2, 2, embed::getIconPixmap("track_op_grip"));
-	}
-	else
-	{
-		p.drawPixmap( 2, 2, embed::getIconPixmap("track_op_grip_c"));
-	}
-}
-
-
-
-
-/*! \brief Clone this track
- *
- */
-void TrackOperationsWidget::cloneTrack()
-{
-	TrackContainerView *tcView = m_trackView->trackContainerView();
-
-	Track *newTrack = m_trackView->getTrack()->clone();
-	TrackView *newTrackView = tcView->createTrackView( newTrack );
-
-	int index = tcView->trackViews().indexOf( m_trackView );
-	int i = tcView->trackViews().size();
-	while ( i != index + 1 )
-	{
-		tcView->moveTrackView( newTrackView, i - 1 );
-		i--;
-	}
-}
-
-
-/*! \brief Clear this track - clears all TCOs from the track */
-void TrackOperationsWidget::clearTrack()
-{
-	Track * t = m_trackView->getTrack();
-	t->addJournalCheckPoint();
-	t->lock();
-	t->deleteTCOs();
-	t->unlock();
-}
-
-
-
-/*! \brief Remove this track from the track list
- *
- */
-void TrackOperationsWidget::removeTrack()
-{
-	emit trackRemovalScheduled( m_trackView );
-}
-
-void TrackOperationsWidget::changeTrackColor()
-{
-	QColor new_color = ColorChooser( this ).withPalette( ColorChooser::Palette::Track )-> \
-		getColor( m_trackView->getTrack()->color() );
-	
-	if( ! new_color.isValid() )
-	{ return; }
-
-	emit colorChanged( new_color );
-	
-	Engine::getSong()->setModified();
-	update();
-}
-
-void TrackOperationsWidget::resetTrackColor()
-{
-	emit colorReset();
-	Engine::getSong()->setModified();
-	update();
-}
-
-void TrackOperationsWidget::randomTrackColor()
-{
-	QColor buffer = ColorChooser::getPalette( ColorChooser::Palette::Track )[ rand() % 48 ];
-
-	emit colorChanged( buffer );
-	Engine::getSong()->setModified();
-	update();
-}
-
-void TrackOperationsWidget::useTrackColor()
-{
-	emit colorParented();
-	Engine::getSong()->setModified();
-}
-
-
-/*! \brief Update the trackOperationsWidget context menu
- *
- *  For all track types, we have the Clone and Remove options.
- *  For instrument-tracks we also offer the MIDI-control-menu
- *  For automation tracks, extra options: turn on/off recording
- *  on all TCOs (same should be added for sample tracks when
- *  sampletrack recording is implemented)
- */
-void TrackOperationsWidget::updateMenu()
-{
-	QMenu * toMenu = m_trackOps->menu();
-	toMenu->clear();
-	toMenu->addAction( embed::getIconPixmap( "edit_copy", 16, 16 ),
-						tr( "Clone this track" ),
-						this, SLOT( cloneTrack() ) );
-	toMenu->addAction( embed::getIconPixmap( "cancel", 16, 16 ),
-						tr( "Remove this track" ),
-						this, SLOT( removeTrack() ) );
-
-	if( ! m_trackView->trackContainerView()->fixedTCOs() )
-	{
-		toMenu->addAction( tr( "Clear this track" ), this, SLOT( clearTrack() ) );
-	}
-	if (QMenu *fxMenu = m_trackView->createFxMenu(tr("FX %1: %2"), tr("Assign to new FX Channel")))
-	{
-		toMenu->addMenu(fxMenu);
-	}
-
-	if (InstrumentTrackView * trackView = dynamic_cast<InstrumentTrackView *>(m_trackView))
-	{
-		toMenu->addSeparator();
-		toMenu->addMenu(trackView->midiMenu());
-	}
-	if( dynamic_cast<AutomationTrackView *>( m_trackView ) )
-	{
-		toMenu->addAction( tr( "Turn all recording on" ), this, SLOT( recordingOn() ) );
-		toMenu->addAction( tr( "Turn all recording off" ), this, SLOT( recordingOff() ) );
-	}
-	
-	toMenu->addSeparator();
-	toMenu->addAction( embed::getIconPixmap( "colorize" ),
-						tr( "Change color" ), this, SLOT( changeTrackColor() ) );
-	toMenu->addAction( embed::getIconPixmap( "colorize" ),
-						tr( "Reset color to default" ), this, SLOT( resetTrackColor() ) );
-	toMenu->addAction( embed::getIconPixmap( "colorize" ),
-						tr( "Set random color" ), this, SLOT( randomTrackColor() ) );
-	toMenu->addSeparator();
-	toMenu->addAction( embed::getIconPixmap( "colorize" ),
-						tr( "Clear clip colors" ), this, SLOT( useTrackColor() ) );
-}
-
-
-void TrackOperationsWidget::toggleRecording( bool on )
-{
-	AutomationTrackView * atv = dynamic_cast<AutomationTrackView *>( m_trackView );
-	if( atv )
-	{
-		for( TrackContentObject * tco : atv->getTrack()->getTCOs() )
-		{
-			AutomationPattern * ap = dynamic_cast<AutomationPattern *>( tco );
-			if( ap ) { ap->setRecording( on ); }
-		}
-		atv->update();
-	}
-}
-
-
-
-void TrackOperationsWidget::recordingOn()
-{
-	toggleRecording( true );
-}
-
-
-void TrackOperationsWidget::recordingOff()
-{
-	toggleRecording( false );
-}
-
-
-// ===========================================================================
-// track
-// ===========================================================================
 
 /*! \brief Create a new (empty) track object
  *
@@ -3069,421 +689,3 @@ BoolModel *Track::getMutedModel()
 	return &m_mutedModel;
 }
 
-
-
-
-
-
-// ===========================================================================
-// trackView
-// ===========================================================================
-
-/*! \brief Create a new track View.
- *
- *  The track View is handles the actual display of the track, including
- *  displaying its various widgets and the track segments.
- *
- *  \param track The track to display.
- *  \param tcv The track Container View for us to be displayed in.
- *  \todo Is my description of these properties correct?
- */
-TrackView::TrackView( Track * track, TrackContainerView * tcv ) :
-	QWidget( tcv->contentWidget() ),   /*!< The Track Container View's content widget. */
-	ModelView( NULL, this ),            /*!< The model view of this track */
-	m_track( track ),                  /*!< The track we're displaying */
-	m_trackContainerView( tcv ),       /*!< The track Container View we're displayed in */
-	m_trackOperationsWidget( this ),    /*!< Our trackOperationsWidget */
-	m_trackSettingsWidget( this ),      /*!< Our trackSettingsWidget */
-	m_trackContentWidget( this ),       /*!< Our trackContentWidget */
-	m_action( NoAction )                /*!< The action we're currently performing */
-{
-	setAutoFillBackground( true );
-	QPalette pal;
-	pal.setColor( backgroundRole(), QColor( 32, 36, 40 ) );
-	setPalette( pal );
-
-	m_trackSettingsWidget.setAutoFillBackground( true );
-
-	QHBoxLayout * layout = new QHBoxLayout( this );
-	layout->setMargin( 0 );
-	layout->setSpacing( 0 );
-	layout->addWidget( &m_trackOperationsWidget );
-	layout->addWidget( &m_trackSettingsWidget );
-	layout->addWidget( &m_trackContentWidget, 1 );
-	setFixedHeight( m_track->getHeight() );
-
-	resizeEvent( NULL );
-
-	setAcceptDrops( true );
-	setAttribute( Qt::WA_DeleteOnClose, true );
-
-
-	connect( m_track, SIGNAL( destroyedTrack() ), this, SLOT( close() ) );
-	connect( m_track,
-		SIGNAL( trackContentObjectAdded( TrackContentObject * ) ),
-			this, SLOT( createTCOView( TrackContentObject * ) ),
-			Qt::QueuedConnection );
-
-	connect( &m_track->m_mutedModel, SIGNAL( dataChanged() ),
-			&m_trackContentWidget, SLOT( update() ) );
-
-	connect(&m_track->m_mutedModel, SIGNAL(dataChanged()),
-			this, SLOT(muteChanged()));
-
-	connect( &m_track->m_soloModel, SIGNAL( dataChanged() ),
-			m_track, SLOT( toggleSolo() ), Qt::DirectConnection );
-			
-	connect( &m_trackOperationsWidget, SIGNAL( colorChanged( QColor & ) ),
-			m_track, SLOT( trackColorChanged( QColor & ) ) );
-			
-	connect( &m_trackOperationsWidget, SIGNAL( colorReset() ),
-			m_track, SLOT( trackColorReset() ) );
-			
-	// create views for already existing TCOs
-	for( Track::tcoVector::iterator it =
-					m_track->m_trackContentObjects.begin();
-			it != m_track->m_trackContentObjects.end(); ++it )
-	{
-		createTCOView( *it );
-	}
-
-	m_trackContainerView->addTrackView( this );
-}
-
-
-
-
-/*! \brief Destroy this track View.
- *
- */
-TrackView::~TrackView()
-{
-}
-
-
-
-
-/*! \brief Resize this track View.
- *
- *  \param re the Resize Event to handle.
- */
-void TrackView::resizeEvent( QResizeEvent * re )
-{
-	if( ConfigManager::inst()->value( "ui",
-					  "compacttrackbuttons" ).toInt() )
-	{
-		m_trackOperationsWidget.setFixedSize( TRACK_OP_WIDTH_COMPACT, height() - 1 );
-		m_trackSettingsWidget.setFixedSize( DEFAULT_SETTINGS_WIDGET_WIDTH_COMPACT, height() - 1 );
-	}
-	else
-	{
-		m_trackOperationsWidget.setFixedSize( TRACK_OP_WIDTH, height() - 1 );
-		m_trackSettingsWidget.setFixedSize( DEFAULT_SETTINGS_WIDGET_WIDTH, height() - 1 );
-	}
-	m_trackContentWidget.setFixedHeight( height() );
-}
-
-
-
-
-/*! \brief Update this track View and all its content objects.
- *
- */
-void TrackView::update()
-{
-	m_trackContentWidget.update();
-	if( !m_trackContainerView->fixedTCOs() )
-	{
-		m_trackContentWidget.changePosition();
-	}
-	QWidget::update();
-}
-
-
-
-
-/*! \brief Create a menu for assigning/creating channels for this track.
- *
- */
-QMenu * TrackView::createFxMenu(QString title, QString newFxLabel)
-{
-	Q_UNUSED(title)
-	Q_UNUSED(newFxLabel)
-	return NULL;
-}
-
-
-
-
-/*! \brief Close this track View.
- *
- */
-bool TrackView::close()
-{
-	m_trackContainerView->removeTrackView( this );
-	return QWidget::close();
-}
-
-
-
-
-/*! \brief Register that the model of this track View has changed.
- *
- */
-void TrackView::modelChanged()
-{
-	m_track = castModel<Track>();
-	assert( m_track != NULL );
-	connect( m_track, SIGNAL( destroyedTrack() ), this, SLOT( close() ) );
-	m_trackOperationsWidget.m_muteBtn->setModel( &m_track->m_mutedModel );
-	m_trackOperationsWidget.m_soloBtn->setModel( &m_track->m_soloModel );
-	ModelView::modelChanged();
-	setFixedHeight( m_track->getHeight() );
-}
-
-
-
-
-/*! \brief Start a drag event on this track View.
- *
- *  \param dee the DragEnterEvent to start.
- */
-void TrackView::dragEnterEvent( QDragEnterEvent * dee )
-{
-	StringPairDrag::processDragEnterEvent( dee, "track_" +
-					QString::number( m_track->type() ) );
-}
-
-
-
-
-/*! \brief Accept a drop event on this track View.
- *
- *  We only accept drop events that are of the same type as this track.
- *  If so, we decode the data from the drop event by just feeding it
- *  back into the engine as a state.
- *
- *  \param de the DropEvent to handle.
- */
-void TrackView::dropEvent( QDropEvent * de )
-{
-	QString type = StringPairDrag::decodeKey( de );
-	QString value = StringPairDrag::decodeValue( de );
-	if( type == ( "track_" + QString::number( m_track->type() ) ) )
-	{
-		// value contains our XML-data so simply create a
-		// DataFile which does the rest for us...
-		DataFile dataFile( value.toUtf8() );
-		Engine::mixer()->requestChangeInModel();
-		m_track->restoreState( dataFile.content().firstChild().toElement() );
-		Engine::mixer()->doneChangeInModel();
-		de->accept();
-	}
-}
-
-
-
-
-/*! \brief Handle a mouse press event on this track View.
- *
- *  If this track container supports rubber band selection, let the
- *  widget handle that and don't bother with any other handling.
- *
- *  If the left mouse button is pressed, we handle two things.  If
- *  SHIFT is pressed, then we resize vertically.  Otherwise we start
- *  the process of moving this track to a new position.
- *
- *  Otherwise we let the widget handle the mouse event as normal.
- *
- *  \param me the MouseEvent to handle.
- */
-void TrackView::mousePressEvent( QMouseEvent * me )
-{
-
-	// If previously dragged too small, restore on shift-leftclick
-	if( height() < DEFAULT_TRACK_HEIGHT &&
-		me->modifiers() & Qt::ShiftModifier &&
-		me->button() == Qt::LeftButton )
-	{
-		setFixedHeight( DEFAULT_TRACK_HEIGHT );
-		m_track->setHeight( DEFAULT_TRACK_HEIGHT );
-	}
-
-
-	int widgetTotal = ConfigManager::inst()->value( "ui",
-							"compacttrackbuttons" ).toInt()==1 ?
-		DEFAULT_SETTINGS_WIDGET_WIDTH_COMPACT + TRACK_OP_WIDTH_COMPACT :
-		DEFAULT_SETTINGS_WIDGET_WIDTH + TRACK_OP_WIDTH;
-	if( m_trackContainerView->allowRubberband() == true  && me->x() > widgetTotal )
-	{
-		QWidget::mousePressEvent( me );
-	}
-	else if( me->button() == Qt::LeftButton )
-	{
-		if( me->modifiers() & Qt::ShiftModifier )
-		{
-			m_action = ResizeTrack;
-			QCursor::setPos( mapToGlobal( QPoint( me->x(),
-								height() ) ) );
-			QCursor c( Qt::SizeVerCursor);
-			QApplication::setOverrideCursor( c );
-		}
-		else
-		{
-			if( me->x()>10 ) // 10 = The width of the grip + 2 pixels to the left and right.
-			{
-				QWidget::mousePressEvent( me );
-				return;
-			}
-
-			m_action = MoveTrack;
-
-			QCursor c( Qt::SizeVerCursor );
-			QApplication::setOverrideCursor( c );
-			// update because in move-mode, all elements in
-			// track-op-widgets are hidden as a visual feedback
-			m_trackOperationsWidget.update();
-		}
-
-		me->accept();
-	}
-	else
-	{
-		QWidget::mousePressEvent( me );
-	}
-}
-
-
-
-
-/*! \brief Handle a mouse move event on this track View.
- *
- *  If this track container supports rubber band selection, let the
- *  widget handle that and don't bother with any other handling.
- *
- *  Otherwise if we've started the move process (from mousePressEvent())
- *  then move ourselves into that position, reordering the track list
- *  with moveTrackViewUp() and moveTrackViewDown() to suit.  We make a
- *  note of this in the undo journal in case the user wants to undo this
- *  move.
- *
- *  Likewise if we've started a resize process, handle this too, making
- *  sure that we never go below the minimum track height.
- *
- *  \param me the MouseEvent to handle.
- */
-void TrackView::mouseMoveEvent( QMouseEvent * me )
-{
-	int widgetTotal = ConfigManager::inst()->value( "ui",
-							"compacttrackbuttons" ).toInt()==1 ?
-		DEFAULT_SETTINGS_WIDGET_WIDTH_COMPACT + TRACK_OP_WIDTH_COMPACT :
-		DEFAULT_SETTINGS_WIDGET_WIDTH + TRACK_OP_WIDTH;
-	if( m_trackContainerView->allowRubberband() == true && me->x() > widgetTotal )
-	{
-		QWidget::mouseMoveEvent( me );
-	}
-	else if( m_action == MoveTrack )
-	{
-		// look which track-widget the mouse-cursor is over
-		const int yPos =
-			m_trackContainerView->contentWidget()->mapFromGlobal( me->globalPos() ).y();
-		const TrackView * trackAtY = m_trackContainerView->trackViewAt( yPos );
-
-		// debug code
-		//	qDebug( "y position %d", yPos );
-
-		// a track-widget not equal to ourself?
-		if( trackAtY != NULL && trackAtY != this )
-		{
-			// then move us up/down there!
-			if( me->y() < 0 )
-			{
-				m_trackContainerView->moveTrackViewUp( this );
-			}
-			else
-			{
-				m_trackContainerView->moveTrackViewDown( this );
-			}
-		}
-	}
-	else if( m_action == ResizeTrack )
-	{
-		setFixedHeight( qMax<int>( me->y(), MINIMAL_TRACK_HEIGHT ) );
-		m_trackContainerView->realignTracks();
-		m_track->setHeight( height() );
-	}
-
-	if( height() < DEFAULT_TRACK_HEIGHT )
-	{
-		ToolTip::add( this, m_track->m_name );
-	}
-}
-
-
-
-/*! \brief Handle a mouse release event on this track View.
- *
- *  \param me the MouseEvent to handle.
- */
-void TrackView::mouseReleaseEvent( QMouseEvent * me )
-{
-	m_action = NoAction;
-	while( QApplication::overrideCursor() != NULL )
-	{
-		QApplication::restoreOverrideCursor();
-	}
-	m_trackOperationsWidget.update();
-
-	QWidget::mouseReleaseEvent( me );
-}
-
-
-
-
-/*! \brief Repaint this track View.
- *
- *  \param pe the PaintEvent to start.
- */
-void TrackView::paintEvent( QPaintEvent * pe )
-{
-	QStyleOption opt;
-	opt.initFrom( this );
-	QPainter p( this );
-	style()->drawPrimitive( QStyle::PE_Widget, &opt, &p, this );
-}
-
-
-
-
-/*! \brief Create a TrackContentObject View in this track View.
- *
- *  \param tco the TrackContentObject to create the view for.
- *  \todo is this a good description for what this method does?
- */
-void TrackView::createTCOView( TrackContentObject * tco )
-{
-	TrackContentObjectView * tv = tco->createView( this );
-	if( tco->getSelectViewOnCreate() == true )
-	{
-		tv->setSelected( true );
-	}
-	tco->selectViewOnCreate( false );
-}
-
-
-
-
-void TrackView::muteChanged()
-{
-	FadeButton * indicator = getActivityIndicator();
-	if (indicator) { setIndicatorMute(indicator, m_track->m_mutedModel.value()); }
-}
-
-
-
-
-void TrackView::setIndicatorMute(FadeButton* indicator, bool muted)
-{
-	QPalette::ColorRole role = muted ? QPalette::Highlight : QPalette::BrightText;
-	indicator->setActiveColor(QApplication::palette().color(QPalette::Active, role));
-}
