+++ src/gui/CMakeLists.txt
@@ -34,6 +34,8 @@ SET(LMMS_SRCS
 	gui/TimeLineWidget.cpp
 	gui/ToolPluginView.cpp
 	gui/TrackContainerView.cpp
+	gui/TrackContentObjectView.cpp
+	gui/TrackView.cpp
 
 	gui/dialogs/FileDialog.cpp
 	gui/dialogs/VersionedSaveDialog.cpp
@@ -97,6 +99,8 @@ SET(LMMS_SRCS
 	gui/widgets/ToolButton.cpp
 	gui/widgets/ToolTip.cpp
 	gui/widgets/TrackLabelButton.cpp
+	gui/widgets/TrackContentWidget.cpp
+	gui/widgets/TrackOperationsWidget.cpp
 	gui/widgets/TrackRenameLineEdit.cpp
 	gui/widgets/StepRecorderWidget.cpp
